+++ src/test/java/roomescape/reservation/controller/ReservationControllerTest.java
@@ -1,106 +1,189 @@
 package roomescape.reservation.controller;
 
-class ReservationControllerTest {
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-//    private FakeReservationService reservationService;
-//    private ReservationController reservationController;
-//
-//    @BeforeEach
-//    void init() {
-//        reservationService = new FakeReservationService();
-//        reservationController = new ReservationController(reservationService);
-//    }
-//
-//    @DisplayName("모든 예약 정보를 가져온다.")
-//    @Test
-//    void test1() {
-//        // given
-//        List<String> names = List.of("꾹", "드라고", "히로");
-//        LocalDate now = LocalDate.now();
-//        LocalTime time = LocalTime.now();
-//        Long timeId = 1L;
-//
-//        reservationService.addReservationTime(timeId, new ReservationTime(timeId, time));
-//
-//        for (String name : names) {
-//            ReservationRequest request = new ReservationRequest(name, now, timeId);
-//            reservationService.save(request);
-//        }
-//
-//        // when
-//        ResponseEntity<List<ReservationResponse>> result = reservationController.readAllReservations();
-//        List<ReservationResponse> body = result.getBody();
-//
-//        // then
-//        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-//
-//        Assertions.assertNotNull(body);
-//
-//        List<String> resultNames = body.stream().map(ReservationResponse::name).toList();
-//        List<LocalDate> resultDates = body.stream().map(ReservationResponse::date).toList();
-//        List<LocalTime> resultTimes = body.stream()
-//                .map(ReservationResponse::time)
-//                .map(ReservationTimeResponse::startAt)
-//                .toList();
-//
-//        assertThat(resultNames).containsExactlyElementsOf(names);
-//
-//        SoftAssertions softly = new SoftAssertions();
-//        for (LocalDate dateTime : resultDates) {
-//            softly.assertThat(dateTime).isEqualTo(now);
-//        }
-//        for (LocalTime resultTime : resultTimes) {
-//            softly.assertThat(resultTime).isEqualTo(time);
-//        }
-//        softly.assertAll();
-//    }
-//
-//    @DisplayName("예약 정보를 추가한다.")
-//    @ParameterizedTest
-//    @CsvSource(value = {"꾹,2025-04-17,1"}, delimiter = ',')
-//    void test2(String name, String date, Long timeId) {
-//        // given
-//        LocalTime localTime = LocalTime.now();
-//        ReservationTime time = new ReservationTime(timeId, localTime);
-//        reservationService.addReservationTime(timeId, time);
-//
-//        LocalDate localDate = LocalDate.parse(date);
-//
-//        ReservationRequest request = new ReservationRequest(name, localDate, timeId);
-//        ReservationTimeResponse reservationTimeResponse = new ReservationTimeResponse(timeId, localTime);
-//        ReservationResponse expected = new ReservationResponse(1L, name, localDate, reservationTimeResponse);
-//
-//        // when
-//        ResponseEntity<ReservationResponse> result = reservationController.save(request);
-//
-//        // then
-//        assertThat(result.getStatusCode()).isEqualTo(HttpStatus.OK);
-//        assertThat(result.getBody()).isEqualTo(expected);
-//    }
-//
-//    @DisplayName("예약 정보를 삭제한다.")
-//    @Test
-//    void test3() {
-//        // given
-//        long id = 1L;
-//        ReservationTime reservationTime = new ReservationTime(id, LocalTime.now());
-//        Reservation reservation = new Reservation(id, "꾹", LocalDate.now(), reservationTime);
-//
-//        reservationService.addReservation(id, reservation);
-//
-//        // when & then
-//        assertThatCode(() -> reservationController.delete(id))
-//                .doesNotThrowAnyException();
-//    }
-//
-//    @DisplayName("삭제할 예약 정보가 없다면 예외를 반환한다")
-//    @Test
-//    void test4() {
-//        // given
-//        long id = 1L;
-//
-//        // when & then
-//        assertThatThrownBy(() -> reservationController.delete(id))
-//                .isInstanceOf(EntityNotFoundException.class);
-//    }
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.lang.reflect.Field;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicLong;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.http.MediaType;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import roomescape.common.exception.EntityNotFoundException;
+import roomescape.reservation.dto.ReservationRequest;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.ReservationTimeResponse;
+import roomescape.reservation.service.ReservationService;
+
+@WebMvcTest(ReservationController.class)
+public class ReservationControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ReservationController reservationController;
+
+    @MockitoBean
+    private ReservationService reservationService;
+
+    @Autowired
+    private ObjectMapper objectMapper;
+
+    @DisplayName("모든 예약 정보를 가져온다.")
+    @Test
+    void test1() throws Exception {
+        // given
+        List<String> names = List.of("꾹", "드라고", "히로");
+        LocalDate now = LocalDate.now();
+        LocalTime time = LocalTime.now();
+        Long timeId = 1L;
+
+        AtomicLong id = new AtomicLong(1L);
+        ReservationTimeResponse timeResponse = new ReservationTimeResponse(timeId, time);
+
+        List<ReservationResponse> responses = names.stream()
+                .map(name -> new ReservationResponse(id.getAndIncrement(), name, now, timeResponse))
+                .toList();
+
+        // when
+        when(reservationService.getAll()).thenReturn(responses);
+
+        // then
+        mockMvc.perform(MockMvcRequestBuilders.get("/reservations"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.length()").value(names.size()));
+    }
+
+    @DisplayName("예약 정보를 추가한다.")
+    @Test
+    void test2() throws Exception {
+        Long reservationId = 1L;
+        Long timeId = 1L;
+        LocalDate date = LocalDate.now();
+        LocalTime time = LocalTime.now();
+        String name = "꾹이";
+
+        ReservationRequest request = new ReservationRequest(name, date, timeId);
+        ReservationTimeResponse timeResponse = new ReservationTimeResponse(timeId, time);
+        ReservationResponse response = new ReservationResponse(reservationId, name, date, timeResponse);
+
+        String requestContent = objectMapper.writeValueAsString(request);
+
+        when(reservationService.save(request)).thenReturn(response);
+
+        mockMvc.perform(MockMvcRequestBuilders.post("/reservations")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestContent))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.id").value(reservationId))
+                .andExpect(jsonPath("$.name").value(name))
+                .andExpect(jsonPath("$.date").value(date.toString()))
+                .andExpect(jsonPath("$.time.id").value(timeId))
+                .andExpect(jsonPath("$.time.startAt").value(formatStartAt(time)));
+    }
+
+    @DisplayName("잘못된 정보가 입력되면 400 상태 코드를 반환한다.")
+    @Test
+    void test3() throws Exception {
+        // given
+        Long timeId = 1L;
+        LocalDate date = LocalDate.now();
+
+        ReservationRequest request = new ReservationRequest(null, date, timeId);
+
+        String requestContent = objectMapper.writeValueAsString(request);
+
+        // when
+        mockMvc.perform(MockMvcRequestBuilders.post("/reservations")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestContent))
+                .andExpect(status().isBadRequest());
+    }
+
+    @DisplayName("존재하지 않는 TimeId면 404를 반환한다.")
+    @Test
+    void test4() throws Exception {
+        // given
+        Long timeId = 1L;
+        LocalDate date = LocalDate.now();
+        String name = "꾹이";
+
+        ReservationRequest request = new ReservationRequest(name, date, timeId);
+
+        String requestContent = objectMapper.writeValueAsString(request);
+
+        // when
+        when(reservationService.save(any())).thenThrow(EntityNotFoundException.class);
+
+        // then
+        mockMvc.perform(MockMvcRequestBuilders.post("/reservations")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(requestContent))
+                .andExpect(status().isNotFound());
+    }
+
+    @DisplayName("예약 정보를 삭제한다.")
+    @Test
+    void test5() throws Exception {
+        // given
+        long reservationId = 1;
+        // when
+        doNothing().when(reservationService).delete(any());
+
+        // then
+        mockMvc.perform(MockMvcRequestBuilders.delete("/reservations/" + reservationId))
+                .andExpect(status().is(204));
+    }
+
+    @DisplayName("삭제할 예약 정보가 존재하지 않는다면 404 에러를 반환한다.")
+    @Test
+    void test6() throws Exception {
+        // given
+        long reservationId = 1;
+
+        // when
+        doThrow(EntityNotFoundException.class).when(reservationService).delete(any());
+
+        // then
+        mockMvc.perform(MockMvcRequestBuilders.delete("/reservations/" + reservationId))
+                .andExpect(status().isNotFound());
+    }
+
+    @DisplayName("컨트롤러 jdbc 분리 테스트")
+    @Test
+    void jdbcBeanTest() {
+        boolean isJdbcTemplateInjected = false;
+
+        for (Field field : reservationController.getClass().getDeclaredFields()) {
+            if (field.getType().equals(JdbcTemplate.class)) {
+                isJdbcTemplateInjected = true;
+                break;
+            }
+        }
+
+        assertThat(isJdbcTemplateInjected).isFalse();
+    }
+
+    private String formatStartAt(LocalTime time) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HH:mm");
+
+        return time.format(formatter);
+    }
+
 }
